+++ src/main/java/com/andersmmg/cityessentials/item/custom/MarkerItem.java
@@ -1,9 +1,23 @@
 package com.andersmmg.cityessentials.item.custom;
 
+import net.minecraft.client.item.TooltipContext;
 import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.text.Text;
+import net.minecraft.util.Formatting;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.List;
 
 public class MarkerItem extends Item {
     public MarkerItem(Item.Settings settings) {
         super(settings);
+    }
+
+    @Override
+    public void appendTooltip(ItemStack stack, @Nullable World world, List<Text> tooltip, TooltipContext context) {
+        super.appendTooltip(stack, world, tooltip, context);
+        tooltip.add(Text.translatable("item.cityessentials.marker.desc").formatted(Formatting.DARK_GRAY));
     }
 }